Fixes for formatting of quantity column
